Add message if error getting website + update updater...

+++ update-marjo21.sh
@@ -3,45 +3,64 @@
 path=`dirname $0`
 cd "$path"
 
-rm bot.pl
-wget https://raw.githubusercontent.com/aaaaadrien/marjo21/master/bot.pl
-chmod +x bot.pl
+me=$0
+arg=$1
 
-rm update-marjo21.sh
-wget https://raw.githubusercontent.com/aaaaadrien/marjo21/master/update-marjo21.sh
-chmod +x update-marjo21.sh
+version="master"
+
+if [[ -z $1 ]]
+	echo "Mise à jour du programme ..."
+	
+	rm update-marjo21.sh
+	wget https://raw.githubusercontent.com/aaaaadrien/marjo21/$version/update-marjo21.sh
+	chmod +x update-marjo21.sh
+
+	./$0 upd
+fi
 
 
-rm public_html/header.php
-wget -P public_html/ https://raw.githubusercontent.com/aaaaadrien/marjo21/master/public_html/header.php
+if [[ -n $1 ]]
+then
 
-rm public_html/footer.php
-wget -P public_html/ https://raw.githubusercontent.com/aaaaadrien/marjo21/master/public_html/footer.php
+	echo "Mise à jour du robot ..."
 
-rm public_html/functions.php
-wget -P public_html/ https://raw.githubusercontent.com/aaaaadrien/marjo21/master/public_html/functions.php
+	rm bot.pl
+	wget https://raw.githubusercontent.com/aaaaadrien/marjo21/$version/bot.pl
+	chmod +x bot.pl
 
-rm public_html/index.php
-wget -P public_html/ https://raw.githubusercontent.com/aaaaadrien/marjo21/master/public_html/index.php
+	echo "Mise à jour du site ..."
 
-rm public_html/about.php
-wget -P public_html/ https://raw.githubusercontent.com/aaaaadrien/marjo21/master/public_html/about.php
+	rm public_html/header.php
+	wget -P public_html/ https://raw.githubusercontent.com/aaaaadrien/marjo21/$version/public_html/header.php
 
-rm public_html/bugs.php
-wget -P public_html/ https://raw.githubusercontent.com/aaaaadrien/marjo21/master/public_html/bugs.php
+	rm public_html/footer.php
+	wget -P public_html/ https://raw.githubusercontent.com/aaaaadrien/marjo21/$version/public_html/footer.php
 
-rm public_html/help.php
-wget -P public_html/ https://raw.githubusercontent.com/aaaaadrien/marjo21/master/public_html/help.php
+	rm public_html/functions.php
+	wget -P public_html/ https://raw.githubusercontent.com/aaaaadrien/marjo21/$version/public_html/functions.php
 
-rm public_html/menu.php
-wget -P public_html/ https://raw.githubusercontent.com/aaaaadrien/marjo21/master/public_html/menu.php
+	rm public_html/index.php
+	wget -P public_html/ https://raw.githubusercontent.com/aaaaadrien/marjo21/$version/public_html/index.php
 
-rm public_html/search.php
-wget -P public_html/ https://raw.githubusercontent.com/aaaaadrien/marjo21/master/public_html/search.php
+	rm public_html/about.php
+	wget -P public_html/ https://raw.githubusercontent.com/aaaaadrien/marjo21/$version/public_html/about.php
 
-rm public_html/stats.php
-wget -P public_html/ https://raw.githubusercontent.com/aaaaadrien/marjo21/master/public_html/stats.php
+	rm public_html/bugs.php
+	wget -P public_html/ https://raw.githubusercontent.com/aaaaadrien/marjo21/$version/public_html/bugs.php
 
-rm public_html/style.css
-wget -P public_html/ https://raw.githubusercontent.com/aaaaadrien/marjo21/master/public_html/style.css
+	rm public_html/help.php
+	wget -P public_html/ https://raw.githubusercontent.com/aaaaadrien/marjo21/$version/public_html/help.php
 
+	rm public_html/menu.php
+	wget -P public_html/ https://raw.githubusercontent.com/aaaaadrien/marjo21/$version/public_html/menu.php
+
+	rm public_html/search.php
+	wget -P public_html/ https://raw.githubusercontent.com/aaaaadrien/marjo21/$version/public_html/search.php
+
+	rm public_html/stats.php
+	wget -P public_html/ https://raw.githubusercontent.com/aaaaadrien/marjo21/$version/public_html/stats.php
+
+	rm public_html/style.css
+	wget -P public_html/ https://raw.githubusercontent.com/aaaaadrien/marjo21/$version/public_html/style.css
+
+fi
